+++ src/oic/__init__.py
@@ -3,9 +3,22 @@ import string
 
 # Since SystemRandom is not available on all systems
 try:
-    import random.SystemRandom as rnd
+    # Python 3.6+, designed for this usecase
+    from secrets import choice
 except ImportError:
-    import random as rnd
+    import random
+    try:
+        # Python 2.4+ if available on the platform
+        _sysrand = random.SystemRandom()
+        choice = _sysrand.choice
+    except AttributeError:
+        # Fallback, really bad
+        import warnings
+        choice = random.choice
+        warnings.warn(
+            "No good random number generator available on this platform. "
+            "Security tokens will be weak and guessable.",
+            RuntimeWarning)
 
 __author__ = 'Roland Hedberg'
 __version__ = '0.12.0'
@@ -27,7 +40,7 @@ def rndstr(size=16):
     :return: string
     """
     _basech = string.ascii_letters + string.digits
-    return "".join([rnd.choice(_basech) for _ in range(size)])
+    return "".join([choice(_basech) for _ in range(size)])
 
 
 BASECH = string.ascii_letters + string.digits + '-._~'
@@ -36,10 +49,10 @@ BASECH = string.ascii_letters + string.digits + '-._~'
 def unreserved(size=64):
     """
     Returns a string of random ascii characters, digits and unreserved
-    characters
+    characters for use as RFC 7636 code verifiers
 
     :param size: The length of the string
     :return: string
     """
 
-    return "".join([rnd.choice(BASECH) for _ in range(size)])
+    return "".join([choice(BASECH) for _ in range(size)])